pyproject.toml
[project]
name = "argus"
version = "0.1.0"
description = "The all-seeing personal assistant"
authors = [
    {name = "Arsene",email = "a.rurangiza@gmail.com"}
]

packages = [
    {include = "argus", from = "src"},
]

readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "openai (>=1.70.0,<2.0.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
mypy = "^1.15.0"
python-dotenv = "^1.1.0"
ruff = "^0.11.4"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["app"]

[tool.ruff]
line-length = 80
indent-width = 4

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
